fix undefined reference errors on mingw
onnx_viewer, menoh_test and vgg16_benchmark refer to symbols under
menoh_impl that are not exported from menoh DLL. This causes link
error and failure of whole build process, which is nuisance for
continuous integration.

As a temporary measure to avoid the problem on mingw, this commit
disable __declspec(dllexport) on mingw. If there are no
__declspec(dllexport), gcc builds a DLL that exports all symbols.